Made test_funk_common suitable for use in multiple compilation units
